resolved comments 2
