+++ data/kadi4mat_json/atmospheric_corrosion_models/din-corrosion-protection-model-iso-9223-compliant.json
@@ -12,57 +12,60 @@
       "type": "dict",
       "value": [
         {
+          "key": "Time exponent",
+          "type": "list",
+          "value": [
+            {
+              "type": "str",
+              "value": "Use DIN recommended time exponents measured from the ISO CORRAG program"
+            },
+            {
+              "type": "str",
+              "value": "Enter manually"
+            }
+          ]
+        },
+        {
           "key": "Corrosion type",
           "type": "list",
           "value": [
             {
               "type": "str",
-              "value": "Insignificant"
+              "value": "Insignificant - Heated rooms with low relative humidity and insignificant air pollution, e.g., offices, schools, museums."
             },
             {
               "type": "str",
-              "value": "Low"
+              "value": "Low - Unheated rooms with fluctuating temperature and relative humidity. Rare condensation and low air pollution, e.g., storage rooms, sports halls."
             },
             {
               "type": "str",
-              "value": "Moderate"
+              "value": "Moderate - Rooms with occasional condensation and moderate air pollution from production processes, e.g., food processing plants, laundries, breweries, dairies."
             },
             {
               "type": "str",
-              "value": "High"
+              "value": "High - Rooms with frequent condensation and high air pollution from production processes, e.g., industrial plants, swimming pools."
             },
             {
               "type": "str",
-              "value": "Very high"
+              "value": "Very high - Rooms with very high frequency of condensation and/or high air pollution from production processes, e.g., mines, industrial cavities, non-ventilated halls in subtropical and tropical climate zones."
             },
             {
               "type": "str",
-              "value": "Extreme"
+              "value": "Extreme - Rooms with nearly constant condensation or extended periods of exposure to extremely high humidity and/or high air pollution from production processes, e.g., non-ventilated halls in humid tropical climates with invasive pollution from outside air, including airborne chlorides and corrosion-promoting dust."
             }
           ]
         },
         {
-          "key": "Corrosion speed",
-          "type": "list",
-          "value": [
-            {
-              "type": "str",
-              "value": "Lower limit"
-            },
-            {
-              "type": "str",
-              "value": "Upper limit"
-            },
-            {
-              "type": "str",
-              "value": "Average"
-            }
-          ]
+          "description": "t indicates the thickness of the metal that corrodes away annually.",
+          "key": "Corrosion rate",
+          "type": "float",
+          "unit": "um year^{-1}",
+          "value": null
         },
         {
           "key": "Temperature",
           "type": "float",
-          "unit": "°C",
+          "unit": "Celcius",
           "value": null
         },
         {
@@ -74,13 +77,13 @@
         {
           "key": "Chloride pollution annual average",
           "type": "float",
-          "unit": "mgCl−dm−2d−1",
+          "unit": "mg dm^{−2} d^{−1}",
           "value": null
         },
         {
           "key": "SO2 pollution annual average",
           "type": "float",
-          "unit": "mgSO2​dm−2d−1",
+          "unit": "mg dm^{−2} d^{−1}",
           "value": null
         }
       ]
@@ -90,14 +93,14 @@
       "type": "dict",
       "value": [
         {
-          "key": "Corrosion Speed",
+          "key": "Corrosion rate",
           "type": "str",
           "value": "r_{\\text {corr }}=1,77 \\cdot P_{\\mathrm{d}}^{0,52} \\cdot \\exp \\left(0,020 \\cdot \\mathrm{RH}+f_{\\mathrm{St}}\\right)+0,102 \\cdot S_{\\mathrm{d}}^{0,62} \\cdot \\exp (0,033 \\cdot \\mathrm{RH}+0,040 \\cdot T)"
         },
         {
           "key": "Temperature Correction Factor",
           "type": "str",
-          "value": "f_{\\mathrm{St}}=0,150 \\cdot(T-10) \\text { wenn } T \\leq 10^{\\circ} \\mathrm{C} \\text {, andernfalls }-0,054 \\cdot(T-10)"
+          "value": "f_{\\mathrm{St}}=0,150 \\cdot(T-10) \\text { wenn } T \\leq 10^{\\circ} \\mathrm{C} \\text {, otherwise }-0,054 \\cdot(T-10)"
         },
         {
           "key": "Material Loss",
@@ -110,27 +113,13 @@
           "value": "D(t>20)=r_{\\text {corr }}\\left[20^b+b\\left(20^{b-1}\\right)(t-20)\\right]"
         }
       ]
-    },
-    {
-      "key": "Time exponent",
-      "type": "list",
-      "value": [
-        {
-          "type": "str",
-          "value": "Use DIN recommeneded time exponents measured from the ISO CORRAG program"
-        },
-        {
-          "type": "str",
-          "value": "Enter manually"
-        }
-      ]
     }
   ],
   "files": [],
   "id": 224,
   "identifier": "din-corrosion-protection-model-iso-9223-compliant",
-  "last_modified": "2024-08-15T11:00:36.821044+00:00",
-  "license": null,
+  "last_modified": "2024-08-26T12:09:22.036734+00:00",
+  "license": "CC-BY-4.0",
   "links": [
     {
       "created_at": "2024-08-14T15:06:38.220656+00:00",
@@ -193,7 +182,7 @@
             "validation": {
               "required": true
             },
-            "value": false
+            "value": true
           },
           {
             "key": "conferenceName",
@@ -229,7 +218,7 @@
         ],
         "id": 71,
         "identifier": "din-en-iso-92232012-05",
-        "last_modified": "2024-08-14T15:06:38.259341+00:00",
+        "last_modified": "2024-08-26T08:58:02.791233+00:00",
         "license": null,
         "tags": [
           "document",
