+++ data/kadi4mat_json/atmospheric_corrosion_models/model_ma2010.json
@@ -12,6 +12,7 @@
       "type": "dict",
       "value": [
         {
+          "description": "The corrosion speed and the exponent used to evaluate the mass loss were evaluated by corrosion experiments performed on low carbon steel (Q235), where samples were exposed to marine environment at two different exposure sites, Wanning city (marine site: 110050 eastern longitude and 18580 northern latitude; SITE 1) and Shenyang city (industrial site: 123240 eastern longitude and 41320 northern latitude; SITE 2).",
           "key": "Corrosion site",
           "type": "list",
           "value": [
@@ -26,6 +27,7 @@
           ]
         },
         {
+          "description": "Distance from the sea.",
           "key": "Distance",
           "type": "float",
           "unit": "m",
@@ -53,9 +55,9 @@
   ],
   "files": [],
   "id": 225,
-  "identifier": "marine-industrial-transition-corrosion-kinetics-mi",
-  "last_modified": "2024-08-15T07:40:09.523159+00:00",
-  "license": null,
+  "identifier": "model_ma2010",
+  "last_modified": "2024-08-26T12:51:02.731654+00:00",
+  "license": "CC-BY-4.0",
   "links": [
     {
       "created_at": "2024-08-14T15:13:28.073793+00:00",
@@ -114,8 +116,8 @@
           }
         ],
         "id": 190,
-        "identifier": "the_atmospheric_corrosion_kinetics_of_low_carbon_s",
-        "last_modified": "2024-08-14T15:13:28.093810+00:00",
+        "identifier": "ma2010",
+        "last_modified": "2024-08-26T12:52:03.042398+00:00",
         "license": null,
         "tags": [],
         "title": "The atmospheric corrosion kinetics of low carbon steel in a tropical marine environment",
