+++ data/kadi4mat_json/atmospheric_corrosion_models/model_soares1999.json
@@ -5,7 +5,7 @@
     "id": 5,
     "orcid": null
   },
-  "description": "The model described in Equation 6 of the document is a non-linear model for corrosion degradation over time. The model is designed to capture the three distinct phases of corrosion typically observed in marine structures:  \n1. **Initial Phase (Protection Phase)**: During this phase, the metal surface is protected by a coating, and there is no corrosion. This phase lasts until the coating loses its effectiveness.     \n2. **Active Corrosion Phase**: Once the coating is compromised, corrosion begins, and the thickness of the plate decreases. This phase is characterized by an exponential increase in corrosion wastage.     \n3. **Stabilization Phase**: Eventually, the corrosion process slows down as the corroded material forms a layer that protects the underlying metal from further exposure to the corrosive environment, effectively halting the corrosion process.  \n\nThe mathematical representation of the model is given by the differential equation:  \n$$ d(t) = d_{\\infty} (1 - e^{-(t - q_c)/q_t}) $$  \n\nHere:\n\n$d(t)$ is the thickness of the corrosion wastage at time $t$. \n\n$d_{\\infty}$ is the long-term thickness of the corrosion wastage. \n\n$q_c$ is the coating life, the time interval during which the coating is effective. \n\n$q_t$ is the transition time, which determines the rate at which the corrosion process progresses after the coating fails.",
+  "description": "The model described in Equation 6 of the paper by Soares and Garbatov (see Soares1999) is a non-linear model for corrosion degradation over time. The model is designed to capture the three distinct phases of corrosion typically observed in marine structures:  \n1. **Initial Phase (Protection Phase)**: During this phase, the metal surface is protected by a coating, and there is no corrosion. This phase lasts until the coating loses its effectiveness.     \n2. **Active Corrosion Phase**: Once the coating is compromised, corrosion begins, and the thickness of the plate decreases. This phase is characterized by an exponential increase in corrosion wastage.     \n3. **Stabilization Phase**: Eventually, the corrosion process slows down as the corroded material forms a layer that protects the underlying metal from further exposure to the corrosive environment, effectively halting the corrosion process.  \n\nThe mathematical representation of the model is given by the differential equation:  \n$$ d(t) = d_{\\infty} (1 - e^{-(t - q_c)/q_t}) $$  \n\nHere:\n\n$d(t)$ is the thickness of the corrosion wastage at time $t$. \n\n$d_{\\infty}$ is the long-term thickness of the corrosion wastage. \n\n$q_c$ is the coating life, the time interval during which the coating is effective. \n\n$q_t$ is the transition time, which determines the rate at which the corrosion process progresses after the coating fails.",
   "extras": [
     {
       "key": "Parameters",
@@ -50,9 +50,9 @@
   ],
   "files": [],
   "id": 227,
-  "identifier": "corrosion-degradation-model-including-coating-effe",
-  "last_modified": "2024-08-15T08:01:55.834569+00:00",
-  "license": null,
+  "identifier": "model_soares1999",
+  "last_modified": "2024-08-26T12:54:40.602156+00:00",
+  "license": "CC-BY-4.0",
   "links": [
     {
       "created_at": "2024-08-15T07:59:38.109998+00:00",
@@ -111,8 +111,8 @@
           }
         ],
         "id": 197,
-        "identifier": "reliability_of_maintained_corrosion_protected_plat",
-        "last_modified": "2024-08-15T07:59:38.116905+00:00",
+        "identifier": "soares1999",
+        "last_modified": "2024-08-26T12:54:48.043136+00:00",
         "license": null,
         "tags": [],
         "title": "Reliability of maintained, corrosion protected plates subjected to non-linear corrosion and compressive loads",
