+++ data/tables/ali2020_tables_table_3.csv
