+++ data/tables/developing_a_risk_assessment_tool_to_predict_the_r_tables_table_8.csv
@@ -19,7 +19,6 @@ Midwest Energy,73,13,15,0.47,215.3,7.07,7.29,17.66,0.0016
 Hallett Dock 5,70,9.7,14,0.40,190.5,7.48,6.78,17.43,0.0017
 Oliver Bridge,81,7.5,10,0.26,203.6,6.71,7.01,19.27,0.0013
 Knife River,45,1.4,3.3,0.12,83.4,8.00,10.85,1.30,0.0029
-Marina,,,,,,,,,
 Two Harbors,43,2.7,3.5,0.17,84.8,7.63,10.28,1.20,0.0011
 Silver Bay,53,1.8,3.5,0.12,83.2,6.83,10.11,1.18,< 0.0008
 Taconite Harbor,380,1.4,3.4,0.01,48.0,7.05,10.38,1.83,0.0015
