+++ web_app/models/IC_model_ali2020.py
@@ -1,8 +1,40 @@
 import pandas as pd
 import streamlit as st
 import numpy as np
-from . immersed_corrosion_models import empirical_prediction_model
+from .immersed_corrosion_models import immersed_corrosion_model
 
+
+'''
+    @article{Ali_2020, title={The empirical prediction of weight change and corrosion rate of low-carbon steel}, 
+    volume={6}, 
+    ISSN={2405-8440}, 
+    url={http://dx.doi.org/10.1016/j.heliyon.2020.e05050}, 
+    DOI={10.1016/j.heliyon.2020.e05050}, 
+    number={9}, 
+    journal={Heliyon}, 
+    publisher={Elsevier BV}, 
+    author={Ali, Nurdin and Fulazzaky, Mohamad Ali}, 
+    year={2020}, 
+    month=sep, 
+    pages={e05050} }
+
+'''
+
+class empirical_prediction_model(immersed_corrosion_model):
+
+    def __init__(self, parameters):
+        immersed_corrosion_model.__init__(self)
+        self.model_name = 'The empirical prediction of weight change and corrosion rate of low-carbon steel'
+        self.article_identifier = ['ali2020']
+        self.steel = "Low carbon steel"
+        self.p = parameters
+
+    
+    def eval_material_loss(self, time):
+        material_loss = (0.00006*self.p['C'] + 0.0008)*time + self.p['b']
+
+        return material_loss
+    
 
 def get_constant_value(nacl_conc, table):
     nacl_concs = np.array(table.iloc[1:, 0].astype(int))
@@ -51,7 +83,7 @@ def display_formulas():
     st.write(r'Mass loss due to corrosion, $W_L [um] = (0.00006C + 0.0008)t + b $ ')
 
 
-def IC_model1(model_identifier):
+def IC_model_ali2020(model_identifier):
     time = st.number_input('Enter duration [years]:', min_value=1.0, max_value=100.0, step=0.1) 
     limits = {
         'C': {'desc': 'Concentration of NaCl', 'lower': 0, 'upper': 5, 'unit': '%w/w'},
